ai_service/requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python
pytest
requests
httpx
